separated setup code; began inner circle drawing
